Update README.md
This project aims to predict the likelihood of diabetes in patients using machine learning techniques based on diagnostic health data. By training models on the Pima Indians Diabetes dataset, we identify whether a patient is likely to have diabetes or not. Early detection can help in better medical intervention and health management.